+++ packages/calcite-components/src/components/list-item-group/list-item-group.tsx
@@ -25,7 +25,7 @@ export class ListItemGroup extends LitElement implements InteractiveComponent {
 
   // #region Public Properties
 
-  /** When `true`, interaction is prevented and the component is displayed with lower opacity. */
+  /** When present, interaction is prevented and the component is displayed with lower opacity. */
   @property({ reflect: true }) disabled = false;
 
   /**