+++ packages/calcite-components/src/components/list-item/list-item.tsx
@@ -51,17 +51,17 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
 
   //#region Private Properties
 
-  private actionsEndEl = createRef<HTMLDivElement>();
+  private actionsEndRef = createRef<HTMLDivElement>();
 
-  private actionsStartEl = createRef<HTMLDivElement>();
+  private actionsStartRef = createRef<HTMLDivElement>();
 
-  private containerEl = createRef<HTMLDivElement>();
+  private containerRef = createRef<HTMLDivElement>();
 
-  private contentEl = createRef<HTMLDivElement>();
+  private contentRef = createRef<HTMLDivElement>();
 
-  private defaultSlotEl = createRef<HTMLSlotElement>();
+  private defaultSlotRef = createRef<HTMLSlotElement>();
 
-  private handleGridEl = createRef<HTMLDivElement>();
+  private handleGridRef = createRef<HTMLDivElement>();
 
   private sortHandleEl: SortHandle["el"];
 
@@ -121,29 +121,29 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
    */
   @property() sortDisabled = false;
 
-  /** When `true`, a close button is added to the component. */
+  /** When present, a close button is added to the component. */
   @property({ reflect: true }) closable = false;
 
-  /** When `true`, hides the component. */
+  /** When present, hides the component. */
   @property({ reflect: true }) closed = false;
 
   /** A description for the component. Displays below the label text. */
   @property() description: string;
 
-  /** When `true`, interaction is prevented and the component is displayed with lower opacity. */
+  /** When present, interaction is prevented and the component is displayed with lower opacity. */
   @property({ reflect: true }) disabled = false;
 
-  /** When `true`, the item is not draggable. */
+  /** When present, the item is not draggable. */
   @property({ reflect: true }) dragDisabled = false;
 
   /**
-   * When `true`, the component displays a draggable button.
+   * When present, the component displays a draggable button.
    *
    * @private
    */
   @property({ reflect: true }) dragHandle = false;
 
-  /** When `true`, expands the component and its contents. */
+  /** When present, expands the component and its contents. */
   @property({ reflect: true }) expanded = false;
 
   /**
@@ -191,7 +191,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
   @property() moveToItems: SortMenuItem[] = [];
 
   /**
-   * When `true`, the item is open to show child components.
+   * When present, the item is open to show child components.
    *
    * @deprecated Use `expanded` prop instead.
    */
@@ -215,7 +215,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
    * */
   @property({ reflect: true }) scale: Scale = "m";
 
-  /** When `true` and the parent `calcite-list`'s `selectionMode` is `"single"`, `"single-persist"', or `"multiple"`, the component is selected. */
+  /** When present and the parent `calcite-list`'s `selectionMode` is `"single"`, `"single-persist"', or `"multiple"`, the component is selected. */
   @property({ reflect: true }) selected = false;
 
   /**
@@ -249,10 +249,10 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
    */
   @property() setSize: number;
 
-  /** When `true`, displays and positions the sort handle. */
+  /** When present, displays and positions the sort handle. */
   @property({ reflect: true }) sortHandleOpen = false;
 
-  /** When `true`, the component's content appears inactive. */
+  /** When present, the component's content appears inactive. */
   @property({ reflect: true }) unavailable = false;
 
   /** The component's value. */
@@ -281,23 +281,18 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
   @method()
   async setFocus(options?: FocusOptions): Promise<void> {
     return this.focusSetter(() => {
-      const {
-        containerEl: { value: containerEl },
-        parentListEl,
-      } = this;
+      const { containerRef, parentListEl } = this;
       const focusIndex = focusMap.get(parentListEl);
 
       if (typeof focusIndex === "number") {
-        const cells = this.getGridCells();
-        if (cells[focusIndex]) {
-          this.focusCell(cells[focusIndex]);
+        const cell = this.getGridCells()[focusIndex];
+        if (cell) {
+          this.focusCell(cell);
           return;
-        } else {
-          return { target: containerEl, includeContainer: true, strategy: "focusable" };
         }
       }
 
-      return { target: containerEl, includeContainer: true, strategy: "focusable" };
+      return { target: containerRef.value, includeContainer: true, strategy: "focusable" };
     }, options);
   }
 
@@ -424,7 +419,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
     }
 
     if (changes.has("displayMode") && this.hasUpdated) {
-      this.handleExpandableChange(this.defaultSlotEl.value);
+      this.handleExpandableChange(this.defaultSlotRef.value);
     }
 
     if (changes.has("expanded") && this.hasUpdated) {
@@ -478,7 +473,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
 
   private handleCalciteInternalListDefaultSlotChanges(event: CustomEvent<void>): void {
     event.stopPropagation();
-    this.handleExpandableChange(this.defaultSlotEl.value);
+    this.handleExpandableChange(this.defaultSlotRef.value);
   }
 
   private setSortHandleEl(el: SortHandle["el"]): void {
@@ -624,10 +619,10 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
 
   private getGridCells(): HTMLDivElement[] {
     return [
-      this.handleGridEl.value,
-      this.actionsStartEl.value,
-      this.contentEl.value,
-      this.actionsEndEl.value,
+      this.handleGridRef.value,
+      this.actionsStartRef.value,
+      this.contentRef.value,
+      this.actionsEndRef.value,
     ].filter((el) => el && !el.hidden);
   }
 
@@ -639,9 +634,9 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
     const { key } = event;
     const composedPath = event.composedPath();
     const {
-      containerEl: { value: containerEl },
-      actionsStartEl: { value: actionsStartEl },
-      actionsEndEl: { value: actionsEndEl },
+      containerRef,
+      actionsStartRef: { value: actionsStartEl },
+      actionsEndRef: { value: actionsEndEl },
       expanded,
       expandable,
     } = this;
@@ -681,7 +676,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
         }
       } else if (currentIndex === 0) {
         this.focusCell(null);
-        containerEl.focus();
+        containerRef.value.focus();
       } else if (cells[currentIndex] && cells[prevIndex]) {
         this.focusCell(cells[prevIndex]);
       }
@@ -787,7 +782,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
         ariaLabel={label}
         class={{ [CSS.dragContainer]: true, [CSS.gridCell]: true }}
         key="drag-handle-container"
-        ref={this.handleGridEl}
+        ref={this.handleGridRef}
         role="gridcell"
       >
         <calcite-sort-handle
@@ -853,7 +848,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
         class={{ [CSS.actionsStart]: true, [CSS.gridCell]: true }}
         hidden={!hasActionsStart}
         key="actions-start-container"
-        ref={this.actionsStartEl}
+        ref={this.actionsStartRef}
         role="gridcell"
       >
         <slot name={SLOTS.actionsStart} onSlotChange={this.handleActionsStartSlotChange} />
@@ -869,7 +864,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
         class={{ [CSS.actionsEnd]: true, [CSS.gridCell]: true }}
         hidden={!(hasActionsEnd || closable)}
         key="actions-end-container"
-        ref={this.actionsEndEl}
+        ref={this.actionsEndRef}
         role="gridcell"
       >
         <slot name={SLOTS.actionsEnd} onSlotChange={this.handleActionsEndSlotChange} />
@@ -961,7 +956,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
           [CSS.nestedContainerExpanded]: this.expandable && this.expanded,
         }}
       >
-        <slot onSlotChange={this.handleDefaultSlotChange} ref={this.defaultSlotEl} />
+        <slot onSlotChange={this.handleDefaultSlotChange} ref={this.defaultSlotRef} />
       </div>
     );
   }
@@ -1009,7 +1004,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
         }}
         key="content-container"
         onClick={this.handleItemClick}
-        ref={this.contentEl}
+        ref={this.contentRef}
         role="gridcell"
       >
         {content}
@@ -1068,7 +1063,7 @@ export class ListItem extends LitElement implements InteractiveComponent, Sortab
             onFocus={this.focusCellNull}
             onFocusIn={this.emitInternalListItemActive}
             onKeyDown={this.handleItemKeyDown}
-            ref={this.containerEl}
+            ref={this.containerRef}
             role="row"
             tabIndex={active ? 0 : -1}
           >